+++ .gitignore
@@ -3,4 +3,4 @@
 *swp
 .cognitoToken
 __pycache__*
-uploads/*
+uploads
