+++ README.md
@@ -1,76 +1,94 @@
 # Blue Core API
 
-API for managing Blue Core resources and workflows using PostgreSQL and Airflow platforms.
+Blue Core API is a REST API for managing Blue Core resources and workflows using PostgreSQL and Airflow platforms. The application relies on Keycloak and Airflow services. To understand the production deployment of Blue Core API you'll want to look at the [Terraform](https://github.com/blue-core-lod/terraform) repository.
 
+Read on for instructions about how to develop this codebase.
 
-## 🗂️ Project structure
-```
-bluecore_store_migrations/
-|  |── env.py
-|  |── versions/
-|
-src/bluecore_api/
-│   │
-│   ├── app/
-│   │   ├── __init__.py
-│   │   ├── main.py
-│   │   ├── config.py
-│   │   
-│   │  
-│   └── models.py
-tests/
-|
-├── pyproject.yaml
-└── README.md
+## 🧰 Prerequisites
+- [uv]
+- [Docker]
+
+## Setup PostgreSQL, Keycloak and Airflow
+
+Blue Core API depends on having PostgreSQL, Airflow and Keycloak running. The easiest way to do this is to clone the [Blue Core Workflows] repository, and start the default configuration:
+
+```shell
+git clone https://github.com/blue-core-lod/bluecore-workflows
+cd bluecore-workflows
+docker compose up 
 ```
 
----
+## Database Models
 
-## 🛠️ Installation
+The Blue Core API depends on database models in the [Blue Core Data Models] to be present and up to date.
 
-### 🧰 Prerequisites
-- [uv](https://github.com/astral-sh/uv)
-- [Docker](https://www.docker.com/)
-- [Blue Core Data Models][BLUECORE_MODELS]
+```shell
+git clone https://github.com/blue-core-lod/bluecore-models
+cd bluecore-models
+uv run alembic upgrade head
+```
  
-### 🔧 Installation instructions
-1.  Run `uv pip install -r requirements.txt`, and follow the instructions that appear.
-2.  Run `docker-compose pull` to pull down all images.
-3.  Clone the [Blue Core Data Models][BLUECORE_MODELS] repository to run the Alembic
-    database migrations.
+## 🔧 Environment
 
----
+Next you will want to clone bluecore_api repository and create a `.env` file that will bring up the application using services that were brought up in the previous step. You should be able to use the following:
+
+```text
+# service uris
+DATABASE_URL="postgresql://airflow:airflow@localhost/bluecore"
+BLUECORE_URL="http://localhost:3000/"
+AIRFLOW_URL="http://localhost:8080/"
+KEYCLOAK_URL="http://localhost:8081/keycloak/"
+
+# keycloak config so blucore_api users can authenticate
+API_KEYCLOAK_CLIENT_ID="bluecore_api"
+API_KEYCLOAK_USER="developer"
+API_KEYCLOAK_PASSWORD="123456"
+
+# credentials so bluecore_api can talk to airflow
+AIRFLOW_WWW_USER_USERNAME="airflow"
+AIRFLOW_WWW_USER_PASSWORD="airflow"
+```
+
+## 💾 Uploads Directory
+
+The bluecore-workflows application has a `uploads` directory in it. You will need to create a symlink to it in your bluecore_api directory. This will allow files uploaded to the API to be available to he Airflow environment.
+
+For example:
+
+```shell
+ln -s ../bluecore-workflows/uploads/ uploads 
+```
 
 ## 🚀 Running the application
-To start all of the supporting services (PostgreSQL, etc.):
-`docker-compose up -d`
 
-The Postgres Docker database will be available on port 5432. When the database is first brought up, 
-the `create-db.sql` script is run that creates a `bluecore` database with a 
-`bluecore_admin` user. 
+Now you are ready to start the application using your new environment file and the fastapi development server, which will auto-load any changes you make to the code:
 
-After the database is up, change directories to the cloned [Blue Core Data Models][BLUECORE_MODELS] and then from that directory run `uv run alembic upgrade head`
-to create the latest database tables and indices for the database.
+```shell
+uv run dotenv run fastapi dev src/bluecore_api/app/main.py --port 3000
+```
 
-**In development**: To start the FastAPI rest server in dev mode:
-1. Run `export DATABASE_URL=postgresql://bluecore_admin:bluecore_admin@localhost/bluecore` to add the needed environmental variable 
-2. Run `export BLUECORE_ENV=https://dev.bcld.info` if you want the change the root for minting URLs that is different from the default of `https://bcld.info`.  
-3. Run the application at *http://localhost:3000*
-`uv run fastapi dev src/bluecore/app/main.py --port 3000`
-4. Look at the API docs at *https://localhost:3000/docs/*
+## Load Data
 
-This is in development mode and code changes will immediately be loaded without having to restart the server.
+If you want to try loading some data you can use the `bluecore` utility:
 
----
+```shell
+uv run bluecore load sample/batch.jsonld 
+```
 
-## 👨‍💻 Developers
+This will load a batch of data to the bluecore_api API, and tell [Blue Core Workflows] to load it.
 
-### 🔐 Bypassing Keycloak 
-To access the API without needing to authenticate with Keycloak: 
-* run `export DEVELOPER_MODE=true` before running the application.
+## HTTP Requests
 
-### 🧹 Linter for Python 
-Bluecore API uses [ruff](https://docs.astral.sh/ruff/)
+To talk directly to the API you will need to pass along a Keycloak access token. During development you can get one by using the included `bluecore` command line tool:
+
+```
+export TOKEN=`uv run bluecore token`
+curl --header "Authorization: Bearer ${TOKEN}" http://localhost:3000/change_documents/instances/page/1
+```
+
+## 🧹 Linting
+
+Bluecore API uses [ruff]
 - `uv run ruff check`
 
 To auto-fix errors in both (where possible):
@@ -81,7 +99,6 @@ Check formatting differences without changing files:
 
 Apply Ruff's code formatting:
 - `uv run ruff format`
-
 
 💡 It's a good idea to run both check and format to catch lint and formatting issues. 
 Github Actions will fail if either check or format fails.
@@ -96,5 +113,8 @@ To run all of the tests:
 To drop into the Python debugger when a test fails add the following parameters to above command:
 - `uv run pytest -s --pdb` 
 
-
-[BLUECORE_MODELS]: https://github.com/blue-core-lod/bluecore-models
+[Blue Core Data Models]: https://github.com/blue-core-lod/bluecore-models
+[Blue Core Workflows]: https://github.com/blue-core-lod/bluecore-workflows
+[ruff]: https://docs.astral.sh/ruff/
+[uv]: https://github.com/astral-sh/uv
+[Docker]: https://www.docker.com/
