--- compose.yml
@@ -1,27 +1,0 @@
-services:
-  bluecore-store:
-    image: pgvector/pgvector:pg17
-    environment:
-      POSTGRES_USER: bluecore_admin
-      POSTGRES_PASSWORD: bluecore_admin
-    ports:
-      - 5432:5432
-    volumes:
-      - postgres-db-volume:/var/lib/postgresql/data
-      - ./create-db.sql:/docker-entrypoint-initdb.d/create_database.sql
-    healthcheck:
-      test: ["CMD", "pg_isready", "-U", "bluecore_admin"]
-      interval: 10s
-      retries: 5
-      start_period: 5s
-    restart: always
-
-  # bluecore-api:
-  #   build: .
-  #   depends_on:
-  #     bluecore-store:
-  #       condition: service_healthy
-    
-volumes:
-  postgres-db-volume:
-
