--- create-db.sql
@@ -1,3 +1,0 @@
-CREATE DATABASE bluecore;
-GRANT ALL PRIVILEGES ON DATABASE bluecore TO bluecore_admin;
-
