+++ pyproject.toml
@@ -13,7 +13,12 @@ dependencies = [
   "bluecore-models>=0.6.0",
   "python-multipart>=0.0.20",
   "fastapi-keycloak-middleware>=1.2.0",
+  "typer>=0.15.4",
+  "python-dotenv>=1.1.0",
 ]
+
+[project.scripts]
+bluecore = "bluecore_api.cli:app"
 
 [dependency-groups]
 dev = [
@@ -30,4 +35,4 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["bluecore_api"]
+packages = ["src/bluecore_api"]
