--- src/bluecore/database.py
@@ -1,15 +1,0 @@
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
-import os
-
-db_url = os.getenv("DATABASE_URL", "")
-engine = create_engine(db_url)
-Session = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-
-
-def get_db():
-    db = Session()
-    try:
-        yield db
-    finally:
-        db.close()
